Move the 'Add exercise contacts' page
